refactor: move `useAppDispatch` and `useAppSelectors` to the `hooks` dirctory.
