+++ src/components/ChatContent/AccumulatedChanges.tsx
@@ -12,11 +12,7 @@ import * as Collapsible from "@radix-ui/react-collapsible";
 import { Chevron } from "../Collapsible";
 import { TruncateLeft } from "../Text";
 import { ScrollArea } from "../ScrollArea";
-import {
-  useAppSelector,
-  useAppDispatch,
-  // useGetManyDiffState,
-} from "../../app/hooks";
+import { useAppSelector, useAppDispatch } from "../../hooks";
 import { useDiffApplyMutation } from "../../hooks";
 import { selectMessages } from "../../features/Chat/chatThread";
 import { createSelector } from "@reduxjs/toolkit";
