+++ src/components/ChatForm/useCommandCompletionAndPreviewFiles.ts
@@ -1,8 +1,7 @@
 import { useState, useEffect, useMemo, useCallback } from "react";
 import { useDebounceCallback } from "usehooks-ts";
 import { Checkboxes } from "./useCheckBoxes";
-import { useAppSelector } from "../../app/hooks";
-import { useHasCaps } from "../../hooks/useHasCaps";
+import { useHasCaps, useAppSelector } from "../../hooks";
 import { addCheckboxValuesToInput } from "./utils";
 import { selectLspPort, selectVecdb } from "../../features/Config/configSlice";
 import {
