+++ src/components/ComboBox/ComboBox.tsx
@@ -8,7 +8,7 @@ import { Popover } from "./Popover";
 import { TruncateLeft } from "../Text";
 import { type DebouncedState } from "usehooks-ts";
 import { CommandCompletionResponse } from "../../services/refact";
-import { useAppSelector } from "../../app/hooks";
+import { useAppSelector } from "../../hooks";
 import { selectSubmitOption } from "../../features/Config/configSlice";
 
 export type ComboBoxProps = {
