+++ src/features/App.tsx
@@ -8,7 +8,7 @@ import { Chat } from "./Chat";
 import { Sidebar } from "../components/Sidebar/Sidebar";
 import { useEventsBusForIDE, useConfig } from "../hooks";
 
-import { useAppDispatch, useAppSelector } from "../app/hooks";
+import { useAppSelector, useAppDispatch } from "../hooks";
 import { FIMDebug } from "./FIM";
 import { store, persistor, RootState } from "../app/store";
 import { Provider } from "react-redux";
