+++ src/features/Chat/chatThread.ts
@@ -14,8 +14,7 @@ import {
 } from "../../services/refact";
 // TODO: update this type
 import type { ChatResponse } from "../../services/refact";
-import { useAppDispatch, useAppSelector } from "../../app/hooks";
-import { useGetToolsQuery } from "../../hooks";
+import { useGetToolsQuery, useAppDispatch, useAppSelector } from "../../hooks";
 import { type AppDispatch, type RootState } from "../../app/store";
 import { parseOrElse } from "../../utils";
 import { formatChatResponse, formatMessagesForLsp } from "./utils";
