+++ src/hooks/useAppDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from "react-redux";
+import type { AppDispatch } from "../app/store";
+
+export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
