+++ src/hooks/useAppSelector.ts
@@ -1,0 +1,4 @@
+import { useSelector } from "react-redux";
+import type { RootState } from "../app/store";
+
+export const useAppSelector = useSelector.withTypes<RootState>();
