+++ src/hooks/useCanUseTools.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { useAppSelector } from "../app/hooks";
+import { useAppSelector } from "./useAppSelector";
 import { useGetToolsQuery } from "./useGetToolsQuery";
 import { useGetCapsQuery } from "./useGetCapsQuery";
 import { selectModel } from "../features/Chat/chatThread";
